Feat. move get function from Portal to PortalTable
