+++ src/components/contents/DetailViewItems/Portal.jsx
@@ -1,22 +1,18 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useAtom, useSetAtom, useAtomValue } from "jotai";
 
 import {
   currentDBIdAtom,
-  currentDocIndexAtom,
   isEditModeAtom,
   relationshipsDataAtom,
   userAtom,
-  docDataAtom,
-  primaryFieldAtom,
   draggingElementAtom,
   elementScaleAtom,
 } from "../../../atoms/atoms";
 
 import fetchData from "../../../utils/axios";
 
-import Loading from "../../shared/Loading";
 import PortalFooter from "./PortalFooter";
 import PortalTable from "./PortalTable";
 import Button from "../../shared/Button";
@@ -31,49 +27,9 @@ function Portal({ index, relationship }) {
   );
 
   const currentDBId = useAtomValue(currentDBIdAtom);
-  const currentDocIndex = useAtomValue(currentDocIndexAtom);
-  const docData = useAtomValue(docDataAtom);
-  const primaryField = useAtomValue(primaryFieldAtom);
 
   const setDraggingElement = useSetAtom(draggingElementAtom);
   const setElementScale = useSetAtom(elementScaleAtom);
-
-  async function getForeignDocuments(relationshipsIndex) {
-    let queryValue = "";
-
-    docData[currentDocIndex]?.fields.forEach(element => {
-      if (primaryField[relationshipsIndex] === element.fieldName) {
-        queryValue = element.fieldValue.trim();
-      }
-    });
-
-    if (relationship._id) {
-      const response = await fetchData(
-        "GET",
-        `users/${userId}/databases/${currentDBId}/relationships/${relationship._id}?primaryFieldValue=${queryValue}`,
-      );
-
-      return response.data;
-    }
-
-    return [];
-  }
-
-  const { data: foreignDocuments, isLoading } = useQuery(
-    ["foreignDocuments1", currentDBId, currentDocIndex, relationship._id],
-    () => getForeignDocuments(index),
-    {
-      enabled:
-        !!userId &&
-        !!currentDBId &&
-        currentDocIndex !== undefined &&
-        !!relationshipsData,
-      refetchOnWindowFocus: false,
-      onFailure: () => {
-        console.log("sending user to errorpage");
-      },
-    },
-  );
 
   async function deleteRelationship(relationshipIndex) {
     await fetchData(
@@ -91,10 +47,6 @@ function Portal({ index, relationship }) {
       console.log("sending user to errorpage");
     },
   });
-
-  if (isLoading) {
-    return <Loading />;
-  }
 
   return (
     <div
@@ -122,11 +74,7 @@ function Portal({ index, relationship }) {
         style={{ height: `${relationship.portalSize}px` }}
         onDoubleClick={() => setIsEditMode(true)}
       >
-        <PortalTable
-          index={index}
-          relationship={relationship}
-          foreignDocuments={foreignDocuments.displayedDocuments}
-        />
+        <PortalTable index={index} relationship={relationship} />
         {isEditMode && (
           <Button
             className="absolute -top-3 -right-3 w-6 rounded-full"
