+++ src/components/contents/DetailViewItems/PortalTable.jsx
@@ -1,10 +1,71 @@
+import { useQuery } from "@tanstack/react-query";
 import { useAtomValue } from "jotai";
+import fetchData from "../../../utils/axios";
 
-import { isEditModeAtom, draggingElementAtom } from "../../../atoms/atoms";
+import {
+  isEditModeAtom,
+  draggingElementAtom,
+  currentDocIndexAtom,
+  docDataAtom,
+  primaryFieldAtom,
+  userAtom,
+  currentDBIdAtom,
+  relationshipsDataAtom,
+} from "../../../atoms/atoms";
 
-function PortalTable({ relationship, foreignDocuments }) {
+import Loading from "../../shared/Loading";
+
+function PortalTable({ index, relationship }) {
+  const { userId } = useAtomValue(userAtom);
+  const currentDBId = useAtomValue(currentDBIdAtom);
+
   const isEditMode = useAtomValue(isEditModeAtom);
   const draggingElement = useAtomValue(draggingElementAtom);
+  const currentDocIndex = useAtomValue(currentDocIndexAtom);
+  const docData = useAtomValue(docDataAtom);
+  const primaryField = useAtomValue(primaryFieldAtom);
+  const relationshipsData = useAtomValue(relationshipsDataAtom);
+
+  async function getForeignDocuments(relationshipsIndex) {
+    let queryValue = "";
+
+    docData[currentDocIndex]?.fields.forEach(element => {
+      if (primaryField[relationshipsIndex] === element.fieldName) {
+        queryValue = element.fieldValue.trim();
+      }
+    });
+
+    if (relationship._id) {
+      const response = await fetchData(
+        "GET",
+        `users/${userId}/databases/${currentDBId}/relationships/${relationship._id}?primaryFieldValue=${queryValue}`,
+      );
+
+      return response.data.foreignDocuments;
+    }
+
+    return [];
+  }
+
+  const { data: foreignDocuments, isLoading } = useQuery(
+    ["foreignDocuments1", currentDBId, currentDocIndex, relationship._id],
+    () => getForeignDocuments(index),
+    {
+      enabled:
+        !!userId &&
+        !!currentDBId &&
+        currentDocIndex !== undefined &&
+        !!relationshipsData,
+      refetchOnWindowFocus: false,
+      onFailure: () => {
+        console.log("sending user to errorpage");
+      },
+    },
+  );
+
+  if (isLoading) {
+    return <Loading />;
+  }
 
   if (!foreignDocuments || foreignDocuments.length === 0) {
     return null;
@@ -47,10 +108,10 @@ function PortalTable({ relationship, foreignDocuments }) {
             })}
           </tr>
         ) : (
-          foreignDocuments.map((element, index) => {
+          foreignDocuments.map((element, fieldIndex) => {
             return (
               <tr
-                key={element.fields[index]._id}
+                key={element.fields[fieldIndex]._id}
                 className="w-[130px] h-10 border border-dark-grey text-center"
               >
                 {element.fields.map(field => {
